Do not cache every readable/executable elf.
This wastes a lot of memory to leave all of these elf files mmap'd in.
This is especially bad when running on host and mmapping in the huge
libartd.so. Read the load base but unmap the elf right away.

Bug: 21372357
(cherry picked from commit 6eab3d438400bc9d6fa2203d4d8d970358fe502d)

Change-Id: I750d47bda8f718046d7444433c2ff7f7fce9af20