Update all mapoff checks to check against 0.
The dlopen of a shared library from an apk results in huge non-zero
map offsets. This can cause failures due to the checks in the code
of the map offset against the elf data.

We don't really support a non-zero offset for an executable map,
so it's easier to assume this is always zero.

Bug: 23192251
Change-Id: Ie4c34ed5c181135ecd3fe7c968b4a424f8eb34b9