Force memory unwinds to use a zero map offset.
The dlopen of a shared library from an apk results in huge non-zero
map offsets. This can cause failures due to the checks in the code
of the map offset against the elf data.

Currently, unwinds only work through a map that is in the process's
memory, so this will cover any attempt to unwind through a shared
library that was loaded from an apk.

Bug: 23348999

(cherry picked from commit e0ab23223a1c3110c9550136b8a119b4c30ec066)

Change-Id: Ic15a959f6a5b68fe50565111d28812f89f2194fa