Add a size when checking local reads/writes.
When doing a local unwind and checking whether a read or a write is valid,
the length of the read/write is not checked. This can lead to crashes if
attempting to read from within a map, but it would cross into out of the
map.

The fix is to check the size of the read/write.

Bug: 25371801
Change-Id: I52556cec06268ce3be5a6fb718147ba62c0a489f