Ignore compressed sections on local unwinds.
When doing local unwinds, attempts to work on compressed sections is
not necessary since this is only called from ART. ART already dumps the
java stacks and that's the only compressed sections that will be used.

It turns out that trying to process the compressed section in local unwinds
can cause problems since the uncompressed section can be large, which
means suddenly a lot more memory being consumed by the process. It also
takes extra time to complete the unwind, which can cause problems when
an ANR occurs, which might trigger a cascade of ANRs since the unwind
is so cpu intensive.

Bug: 27391690
Change-Id: I98be5754cc4de4071616db334dc0c9862d3a2504
(cherry picked from commit 6292bbbc3f678648fb5a1ec8222d33f6a5bb992a)