Do not try non-dwarf local unwinds for x86.
The frame-chain unwind code in libunwind never really works and it
has a high probability of trying to access unmapped memory. When
doing a remote unwind, this will simply result in the unwind stopping.
When doing a local unwind, this can lead to crashes. Disable this
method for local unwinds.

Bug: 26807329
Change-Id: Iaa2915a38cf12c3e7ff02537a76369be41c60dd7