Use zero offset when getting bias from memory.
The code to get the bias was incorrectly using the passed in map offset
when reading it from memory. This doesn't work since the offset is the
offset into the apk. Change to always check for a zero in this case.

Bug: 28685297
(cherry picked from commit c3290dfdd3ef01a52a544b0ee7e37663befb6a3b)

Change-Id: I6d036bae1bb1ff933e399dc238c0c9233b25f317