Save the pc for aarch64 current unwind.
I had forgotten that I never properly recorded the pc correctly for
aarch64. I was simply using the lr value and assuming it would be
good enough. In many cases it was, but if the sp was not the same,
then an unwind of the currently executing thread would go off into
the weeds.

Actually, record the pc using the adr instruction.

Test: Run backtrace unit tests on angler.

Change-Id: I2df2dea77608f87c82bcf086dce75be1d34e86d1